+++ Changelog.md
@@ -1,3 +1,22 @@
+# 2.3.0
+
+## Common
+
+- Remove JSON datasheet format
+- Add YAML datasheet format
+- Remove `cace_compat.py`, as JSON is not supported anymore
+- Remove standalone `cace_read.py` and `cace_write.py`
+
+## GUI
+
+- Remove `JSON` in the file picker
+- Remove `YAML` in the file picker
+- Remove bit-rotted `load_results`
+
+## CLI
+
+- Remove `--json` argument
+
 # 2.2.7
 
 ## Common